refactor(config): own built-in keymap presets
